requirements.txt
yfinance>=0.2.36
pandas>=2.0.0
pandas-ta>=0.3.14b
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
hmmlearn>=0.3.0
arch>=6.2.0
transformers>=4.35.0
torch>=2.0.0
feedparser>=6.0.10
requests>=2.31.0
optuna>=3.4.0
plotly>=5.18.0
sqlalchemy>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0
vectorbt>=0.26.1
APScheduler>=3.10.4
python-dotenv>=1.0.0
httpx>=0.25.0
pgmpy>=0.1.24
pymc>=5.0.0
